add is.ts for isFunc
