+++ packages/verify/src/vpVerify.ts
@@ -15,7 +15,7 @@ import { assert } from '@polkadot/util';
 
 import { isSameUri } from '@zcloak/did/utils';
 import { hashDigests } from '@zcloak/vc';
-import { isVP } from '@zcloak/vc/utils';
+import { isVP } from '@zcloak/vc/is';
 
 import { proofVerify } from './proofVerify';
 import { vcVerify, vcVerifyDigest } from './vcVerify';
